+++ BadMedicine/Configuration/Config.cs
@@ -1,0 +1,7 @@
+﻿namespace BadMedicine.Configuration
+{
+    class Config
+    {
+        public TargetDatabase Database { get;set; }
+    }
+}